requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.2
h11==0.14.0
idna==3.4
numpy==1.26.1
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.23.1
trio-websocket==0.11.1
tzdata==2023.3
urllib3==2.0.7
wsproto==1.2.0
